fix: added current theme color reletion to svg, refactored code according to comments
